+++ discovery-server/src/main/java/app/metatron/discovery/domain/scheduling/common/OAuthCacheOutJob.java
@@ -32,6 +32,8 @@ import app.metatron.discovery.common.oauth.token.cache.AccessTokenCacheRepositor
 import app.metatron.discovery.common.oauth.token.cache.RefreshTokenCacheRepository;
 import app.metatron.discovery.common.oauth.token.cache.WhitelistTokenCacheRepository;
 
+import static app.metatron.discovery.common.oauth.token.cache.AccessTokenCacheRepository.REFRESH_TO_ACCESS;
+
 @Component
 @Transactional(readOnly = true, isolation = Isolation.READ_UNCOMMITTED)
 public class OAuthCacheOutJob extends QuartzJobBean {
@@ -59,10 +61,14 @@ public class OAuthCacheOutJob extends QuartzJobBean {
     LOGGER.info("## Start batch job for checking expired oauth access token.");
     SpringCache accessTokenCache = (SpringCache) cacheManager.getCache("access-token-cache");
     accessTokenCache.getNativeCache().forEach((key, cachedAccessToken) -> {
-      Date expiration = ((AccessTokenCacheRepository.CachedAccessToken) cachedAccessToken).getExpiration();
-      if(expiration.getTime() < System.currentTimeMillis()){
-        LOGGER.debug("Access Token expired : {}", key);
-        accessTokenCacheRepository.removeAccessToken(key.toString());
+      String tokenKey = key.toString();
+      if(!tokenKey.startsWith(REFRESH_TO_ACCESS)){
+        Date expiration = ((AccessTokenCacheRepository.CachedAccessToken) cachedAccessToken).getExpiration();
+        if(expiration.getTime() < System.currentTimeMillis()){
+          LOGGER.debug("Access Token expired : {}", key);
+          accessTokenCacheRepository.removeAccessToken(key.toString());
+          whitelistTokenCacheRepository.removeWhitelistTokenByCachedAccessToken((AccessTokenCacheRepository.CachedAccessToken) cachedAccessToken);
+        }
       }
     });
     LOGGER.info("## End batch job for checking expired oauth refresh token.");
@@ -74,6 +80,7 @@ public class OAuthCacheOutJob extends QuartzJobBean {
       if(expiration.getTime() < System.currentTimeMillis()){
         LOGGER.debug("Refresh Token expired : {}", key);
         refreshTokenCacheRepository.removeRefreshToken(key.toString());
+        accessTokenCacheRepository.removeAccessTokenByRefreshToken(key.toString());
       }
     });
     LOGGER.info("## End batch job for checking expired oauth refresh token.");
